+++ R/badge_download.R
@@ -9,6 +9,7 @@
 #'
 #' @return character
 #' @import glue
+#' @import cli
 #' @export
 #' @examples
 #' add2blog::badge_download(
@@ -21,7 +22,10 @@ badge_download <- function(
     ,your_github_repo = 'add2prep'
     ,type = c('total')
 ){
-    glue('  [![GitHub All Releases](https://img.shields.io/github/downloads/{your_github_name}/{your_github_repo}/{type}.svg)](https://github.com/JiaxiangBU/add2prep)') %>%
-        print
-    glue('<<<Copy it above nto your README.md>>>>')
+    text = glue('  [![GitHub All Releases](https://img.shields.io/github/downloads/{your_github_name}/{your_github_repo}/{type}.svg)](https://github.com/JiaxiangBU/add2prep)')
+    text %>% cat
+    clipr::write_clip(text)
+    cat('\n\n')
+    glue::glue('This text is on your system clipboard, copy it above into your README.md') %>%
+        cli::rule(center = ., col = "red")
 }